requirements_phoenix.txt
customtkinter>=5.0
requests>=2.20
cryptography>=3.0
Pillow>=9.0
plyer>=2.0
psutil>=5.0
yt-dlp>=2024.05.27 # Ili najnovija verzija